Refactor/abstract GitHub task (#3)
* Add AbstractGithubTask and refactor GithubPublish

The `GithubPublish` task used to extend `Copy` to leverage the copy
logic for asset preparation during task execution. To create an
`abstract` base task I needed to get rid of this inheritance. The task
implements `CopySpecs` sub interfaces (`CopySourceSpec`,
`PatternFilterable`) needed to function as before. All github
authentication related logic and properties are located at
`AbstractGithubTask` This makes it possible to create other tasks with
the basic authentication setup.

* Change propertyName repository to repositoryName

This change was needed to have a clear naming scheme when handling with
github repository names and API objects

* Add generic Github gradle task type

This task allows the user to issue github API calls to the provided
repository. It also exposes the `GHClient` object to do other API
related jobs.

**create repos**
```
customGithubTask {
    doLast {
        def builder =client.createRepository("Repo")
        builder.description("description")
        builder.autoInit(false)
        builder.licenseTemplate('MIT')
        builder.private_(false)
        builder.issues(false)
        builder.wiki(false)
        builder.create()
    }
}
```

**update files**
```
customGithubTask {
    doLast {
        def content = repository.getFileContent("$file")
        content.update("$updatedContent", "update release notes")
    }
}
```

* Update README.md with Github task examples

* Restructe delegates and add test specs for coverage