Fix draft github release when release has no assets
Description
===========

When creating a release with no assets configured to upload, the release
will be a draft only release because the last steps to mark the release
as done are not executed. This change will add more assertions to the
specs to make sure that a complete release is never a `draft` and
changes the logic to create a non draft release right from the start
when no assets are provided.

Changes
=======

![FIX] release logic and mark release as non draft in all cases
![ADD] `!isDraft()` assertions to test specs