Update development and support gradle version to 4.10 (#21)
* Update development and support gradle version to 4.10

Description
===========

This patch updates the used gradle version for development to `4.10`.
This concludes the support of the gradle `4.x` version range as `4.10`
is the last version before `5.0`

Changes
=======

![UPDATE] gradle development/support version to `4.10`

* Gradle >= 4.8 changed the output stream for errors

Errors that normally were pushed through `stderr` are now pushed through
`stdout`. This is a small unitily fix to check for error messages in a
backwardscompatible way.