Add update release feature (#26)
Description
===========

For some publish workflows it would be nice to execute multiple github
publish tasks on multiple gradle runs to support multi platform releases
or integration of third party publish plugins which have their own
workflows. Up until know the only way to publish artifacts from multiple
platforms was to build them first and execute one build to create the
release and upload the artifacts in one run.
With this patch it is finally possible to configure the `GithubPublish`
task type to `create`,`update` or `createOrUpdate` a release.

See a potential pipeline:

```
                                                                       builds in parallel
                                                                       +----------------+

                                                    +------------------------+    +-------------------+
                                                 +-->  build macOS binary    +----> publish binaries  +-+
                                                 |  +------------------------+    +-------------------+ |
                                                 |                                                      |
                                                 |                                                      |
                                                 |                                                      |
+-------------------+     +-------------------+  |  +------------------------+    +-------------------+ |   +-------------------+
|     run tests     +----->   draft release   +----^+  build windows binary  +----> publish binaries  +----->  publish release  |
+-------------------+     +-------------------+  |  +------------------------+    +-------------------+ |   +-------------------+
                                                 |                                                      |
                                                 |                                                      |
                                                 |                                                      |
                                                 |  +------------------------+    +-------------------+ |
                                                 +-->  build linux binary    +----> publish binaries  +-+
                                                    +------------------------+    +-------------------+

```

_the `draft release` step could also be done by the first `publish binaries` step.
First one drafts the release_

This patch includes some big changes in the `GithubPublish` task. I
decided to implement the `update`/`create` handling inside the already
know `GithubPublish` class. The steps of creating a release and updating
it are not much different. The task already had some basic update
functionality to move it from `draft` to `published` state.

I increased the error handling code and release rollback logic, in order to deal
with potential error cases during update of the release meta data or release assets.

Usage
-----

If you want to use this new functionality you have to instruct the
gradle task `githubPublish` or your own custom task if you created one
which publish method to use. There are three options:

| method         | description                                                                   |
| -------------- | ----------------------------------------------------------------------------- |
| create         | creates a new release and fails if one already exists. _This is the default__ |
| update         | updates a release and fails if it doesn't exist.                              |
| updateOrCreate | updates a release or creates it if not created already                        |

> Note: As the common key to find releases at github this plugin uses the
`tagName` property. Github allows to update the `tagName` of a release
but this plugin doesn't support this!

The `GithubPublish` task contains a new property `publishMethod` which
can be set in the __build.gradle__

> Note: There is no default mapping from properties or environment
variables available at this time. The `publishMethod` property can be
used with a closure. This part will be rewritten when the `Provider` API in the future.

**build.gradle**
```
task testPublish(type:wooga.gradle.github.publish.tasks.GithubPublish) {
   from "fileToPublish"
   tagName = "$tagName"
   publishMethod = "update"
}
```

Changes
=======

![ADD] update releases feature