Refactor plugin lazy configuration
Description
===========

The plugin as it is now relies heavily on `convention mapping`, a
technique which is not part of the public gradle API. This patch removes
all usages of `convention mapping` and uses the [`Provider`] API instead.
As of this moment this API is still in _incubation_ phase but is a
better alternative aready. This patch also means a breaking change not
only because of internal API changes but also because the new API in the
form of usage was introduced with gradle > 4.7

Changes
=======

![IMPROVE] use [`Provider`] API
![REMOVE] convention mapping
![BREAK] support with gradle < 4.7

[`Provider`]:    https://docs.gradle.org/current/userguide/lazy_configuration.html