Improve publish body input parameter handling (#36)
Description
===========

This patch improves the handling of body property in `GithubPublish`.
With this change it is possible to provide a `File` or `Task` as the
input for the body property. A `File` object will be evaluated and read
during runtime and set as input file.

If the body value is of type `Task`, the task outputs will be evaluated
at runtime. The task will fail if the provided task has to no or too
many output files configured.

Changes
=======

![IMPROVE] publish body input parameter