Fix NullPointerException during exception throw (#59)
* Fix default value for targetCommitish property

Description
===========

This plugin assumed that the best default value for the targetCommitish
property should be `master`. That is wrong for multiple reasons:

1. it's not written in stone what the default branch for a github repo is
2. github themselve changed the default value to `main` in summer 2020

This patch changes this optional value to `null` so that github will use
the default branch name without this plugin setting any value
beforehand.

Changes
=======

* ![FIX] default value for `targetCommitish` property

* Fix NullPointerException during exception throw

Description
===========

This is a potential copy paste error. The object `release` is null at
the time when the publish step might throw an error. The patch simply
takes the tag name from the internal property instead of the release
object.

Changes
=======

* ![FIX] `NullPointerException` during exception throw